requirements.txt
annotated-types==0.6.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dacite==1.8.1
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
email_validator==2.1.2
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fonttools==4.51.0
git-filter-repo==2.38.0
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ImageHash==4.3.1
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
multimethod==1.11.2
-e git+ssh://git@github.com/ahmedaao/news-recommender-systems.git@182067af10105a067c0bca97f1a2102084a23295#egg=myproject
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
orjson==3.10.5
packaging==24.0
pandas==2.2.2
pandas-profiling==3.6.6
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
phik==0.12.4
pillow==10.3.0
platformdirs==4.2.1
plotly==5.22.0
polars==0.20.31
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.31.0
rich==13.7.1
scikit-learn==1.5.0
scikit-surprise==1.1.4
scipy==1.13.0
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.2
tenacity==8.3.0
threadpoolctl==3.5.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typeguard==4.2.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
visions==0.7.6
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
widgetsnbextension==4.0.10
wordcloud==1.9.3
ydata-profiling==4.8.3
